fix: out of range value para presupuesto
Su razon era el uso de INT en la DB, cuyo limite es de 2147483647
negativo y positivo. Se cambio el tipo a BIGINT y se agregaron
validaciones para que el valor ingresado no supere los nuevos limites.

+++ app/schema.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS proyecto(
     descripcion TINYTEXT,
     es_publico BOOLEAN NOT NULL DEFAULT false,
     activo BOOLEAN NOT NULL DEFAULT true,
-    presupuesto INT NOT NULL DEFAULT -1,
+    presupuesto BIGINT NOT NULL DEFAULT -1,
     fecha_inicio DATE NOT NULL DEFAULT (CURRENT_DATE()),
     fecha_finalizacion DATE NOT NULL DEFAULT '1000-01-01'
 );
